moved continuum stuff to transform
