+++ libpysat/__init__.py
@@ -1,12 +1,15 @@
 __version__ = "0.1.0"
 
 from libpysat import *
-from libpysat.data.spectra import Spectra
-from libpysat.data.spectrum import Spectrum
-from libpysat.data.hcube import HCube
-from libpysat.data.io import *
 
 from libpysat import derived
 from libpysat import transform
 from libpysat import regression
 from libpysat import data
+
+from libpysat.data.spectra import Spectra
+from libpysat.data.spectrum import Spectrum
+from libpysat.data.hcube import HCube
+
+from libpysat.data import io
+from libpysat.transform import continuum
